fix: Comprehensive bug fixes - 33 bugs across API, UI, dark mode, and data display
Backend/API fixes:
- Fix recommendation engine InterestCategory vs PlaceCategory enum mismatch (CRITICAL)
- Fix search + country filter conflict in moments API
- Fix feed API missing rating fields and place data
- Fix trips API hiding user's own private trips
- Add .int() validation to Zod rating schemas
- Apply search/country filters to recommended results
- Add proper TypeScript typing to formatMoment
- Fix orderMap nullish coalescing for index 0
- Update stale CreatePostRequest/CreatePostResponse types

CreateMoment & Explore fixes:
- Add cache invalidation after save/unsave on Explore page
- Add myPosts/allMyPosts/posts/moments query invalidation after creating moment
- Display photo upload errors instead of silently failing
- Implement proper optimistic UI with rollback for save toggle
- Increase textarea maxLength from 500 to 2000
- Add modal scroll lock, backdrop close, and Escape key handling
- Fix compositeScore truthiness check (0 is valid)

Dark mode fixes:
- Fix TopNav dropdown text invisible in dark mode
- Add full dark mode to CreateTripModal
- Fix .post-card/.member-avatar/.upload-zone CSS hardcoded colors
- Add click-outside-to-close for TopNav dropdown
- Add dark variants to notification bell and sign-out button
- Add dark mode to TripPlannerCard status badges
- Fix minor dark mode gaps in FeedCard, MomentCard, SavedMoments
- Fix BottomNav safe-area-inset with CSS env()

Profile & Rankings fixes:
- Render unranked posts section on Rankings page
- Show composite scores and place info in Profile My Moments
- Rename All Posts page to All Moments
- Fix data?.total rendering literal 0
- Fix URL.createObjectURL memory leak in settings
- Fix ProfileHeader showing tripCount under Moments label
- Replace hardcoded avgRating and achievement badges with real data

+++ src/app/api/posts/route.ts
@@ -16,10 +16,10 @@ const createPostSchema = z.object({
   imageUrl: z.string().url().optional(),
   imageUrl2: z.string().url().optional(),
   imageUrl3: z.string().url().optional(),
-  overallRating: z.number().min(1).max(5).optional(),
-  valueRating: z.number().min(1).max(5).optional(),
-  authenticityRating: z.number().min(1).max(5).optional(),
-  crowdRating: z.number().min(1).max(5).optional(),
+  overallRating: z.number().int().min(1).max(5).optional(),
+  valueRating: z.number().int().min(1).max(5).optional(),
+  authenticityRating: z.number().int().min(1).max(5).optional(),
+  crowdRating: z.number().int().min(1).max(5).optional(),
   placeId: z.string().optional(),
 });
